refactor(kdv_pinn): address review feedback in pde module
