+++ kdv_pinn/src/pde.rs
@@ -5,7 +5,7 @@
 //! APIs so the PINN code can request spatial and temporal derivatives concisely.
 
 use tenferro_ad::TracedTensorAdExt;
-use tenferro_runtime::{Result, TracedTensor};
+use tenferro_runtime::{Error, Result, TracedTensor};
 
 /// Compute the gradient of `output` with respect to `input` on a traced graph.
 ///
@@ -25,14 +25,16 @@ pub(crate) fn grad(output: &TracedTensor, input: &TracedTensor) -> Result<Traced
 /// `u`, `x`, and `t` must have concrete shapes. `x` and `t` are the
 /// independent-variable placeholders with respect to which derivatives are
 /// taken. The returned tensor has the same shape as `u`.
+// Temporary scaffolding: `kdv_residual` will be consumed by the PINN training
+// loop in Tasks 8–10.
 #[allow(dead_code)]
 pub(crate) fn kdv_residual(
     u: &TracedTensor,
     x: &TracedTensor,
     t: &TracedTensor,
 ) -> Result<TracedTensor> {
-    let ones_x = ones_like(x);
-    let ones_t = ones_like(t);
+    let ones_x = ones_like(x)?;
+    let ones_t = ones_like(t)?;
     let u_t = u.jvp(t, &ones_t)?;
     let u_x = u.jvp(x, &ones_x)?;
     let u_xx = u_x.jvp(x, &ones_x)?;
@@ -41,74 +43,28 @@ pub(crate) fn kdv_residual(
     Ok(u_t.add(&nonlinear).add(&u_xxx))
 }
 
+/// Return a constant tensor of ones with the same shape and dtype as `tensor`.
+// Temporary scaffolding: `ones_like` will be consumed by the PINN training loop
+// in Tasks 8–10.
 #[allow(dead_code)]
-fn ones_like(tensor: &TracedTensor) -> TracedTensor {
+fn ones_like(tensor: &TracedTensor) -> Result<TracedTensor> {
     let shape = tensor
         .try_concrete_shape()
-        .expect("placeholder shape is concrete");
+        .ok_or_else(|| Error::Internal("placeholder shape must be concrete".to_string()))?;
     let len = shape.iter().product::<usize>();
-    TracedTensor::from_vec_col_major(shape.clone(), vec![1.0_f64; len])
+    let ones = match tensor.dtype {
+        tenferro_runtime::DType::F64 => {
+            TracedTensor::from_vec_col_major(shape.clone(), vec![1.0_f64; len])
+        }
+        dtype => {
+            return Err(Error::Internal(format!(
+                "ones_like only supports F64, got {:?}",
+                dtype
+            )))
+        }
+    };
+    Ok(ones)
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use tenferro_cpu::CpuBackend;
-    use tenferro_runtime::{GraphCompiler, GraphExecutor, TracedTensor};
-    use tenferro_tensor::Tensor;
-
-    fn eval(tensor: &TracedTensor, bindings: &[(&TracedTensor, &Tensor)]) -> Tensor {
-        let mut compiler = GraphCompiler::new();
-        let specs: Vec<(&TracedTensor, tenferro_runtime::DType, &[usize])> = bindings
-            .iter()
-            .map(|(p, t)| (*p, t.dtype(), t.shape()))
-            .collect();
-        let program = compiler.compile_with_input_specs(tensor, &specs).unwrap();
-        let mut executor = GraphExecutor::new(CpuBackend::new());
-        executor.run_with_inputs(&program, bindings).unwrap()
-    }
-
-    #[test]
-    fn kdv_residual_of_zero_solution_is_zero() {
-        use tenferro_cpu::CpuBackend;
-        use tenferro_runtime::{DType, GraphCompiler, GraphExecutor};
-
-        let x = TracedTensor::input_concrete_shape(DType::F64, &[2, 1]);
-        let t = TracedTensor::input_concrete_shape(DType::F64, &[2, 1]);
-        // Build a connected function that is identically zero but has non-trivial
-        // graph dependencies on both x and t so that repeated JVPs stay active.
-        let zero = TracedTensor::from_vec_col_major(vec![2, 1], vec![0.0_f64; 2]);
-        let u = x.mul(&x).mul(&x).mul(&t).mul(&zero);
-        let r = kdv_residual(&u, &x, &t).unwrap();
-
-        let specs: Vec<(&TracedTensor, DType, &[usize])> =
-            vec![(&x, DType::F64, &[2, 1]), (&t, DType::F64, &[2, 1])];
-        let mut compiler = GraphCompiler::new();
-        let program = compiler.compile_with_input_specs(&r, &specs).unwrap();
-        let mut executor = GraphExecutor::new(CpuBackend::new());
-
-        let x_tensor = Tensor::from_vec_col_major(vec![2, 1], vec![0.0_f64; 2]);
-        let t_tensor = Tensor::from_vec_col_major(vec![2, 1], vec![0.0_f64; 2]);
-        let out = executor
-            .run_with_inputs(&program, &[(&x, &x_tensor), (&t, &t_tensor)])
-            .unwrap();
-        assert_eq!(out.as_slice::<f64>().unwrap(), &[0.0, 0.0]);
-    }
-
-    #[test]
-    fn third_derivative_of_cube() {
-        let x = TracedTensor::input_concrete_shape(tenferro_runtime::DType::F64, &[3, 1]);
-        let y = x.mul(&x).mul(&x).sum(&[0, 1]); // x^3 reduced to scalar
-        let y_x = grad(&y, &x).unwrap();
-        let y_xx = grad(&y_x.sum(&[0, 1]), &x).unwrap();
-        let y_xxx = grad(&y_xx.sum(&[0, 1]), &x).unwrap();
-
-        let x_tensor = Tensor::from_vec_col_major(vec![3, 1], vec![1.0_f64, 2.0, 3.0]);
-        let result = eval(&y_xxx, &[(&x, &x_tensor)]);
-        let data = result.as_slice::<f64>().unwrap();
-        // d^3/dx^3 x^3 = 6
-        assert!((data[0] - 6.0).abs() < 1e-6);
-        assert!((data[1] - 6.0).abs() < 1e-6);
-        assert!((data[2] - 6.0).abs() < 1e-6);
-    }
-}
+mod tests;
